DATA: add database files
Added the necessary files for the DB to talk to the app. Went with the Repo approach as opposed to the raw DbContext approach mentioned in the videos. Repo only has a Tours queryset currently.

Added a SeedData class to get us all on the same page. SeedData will create Tours occuring on the hour, from 8 AM to 8PM, everyday from 13 Mar 2021 to 27 Mar 2021.