REF: rework all models and view models
Renamed TourParty to just Party. Added navigation properties to Party on Tour. Removed a useless method from Tour (AddParty). The SignUpViewModel now contains all data for a Party and the TourId, but not the Tour. Changing the UserForm view to use a Party was throwing a required error due to Tour not being set.