REF: update controller
Due to model changes, refactored the controller class and how it interacted with the user.

Added logging on each action to make debugging easier.

GET SignUp will now return only those Tours that are in the future and don't have a Party associated with them.

POST SignUp initializes a new SignUpViewModel with the proper TourId in the ViewModel.

GET UserForm creates a new SignUpViewModel with the proper TourId as passed in the GET parameters. This action should never be called.

POST UserForm takes the data from the SignUpViewModel, which mirrors the Party object, and creates a new Party using the data. It links the Tour to the Party properly.

Appointments properly returns a list of all Tours that do not have a party associated with them.